docs/server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
pyjwt==2.8.0
dashscope==1.14.0
gunicorn==21.2.0
openai==1.12.0
pdfplumber==0.10.0
beautifulsoup4==4.12.0
edge-tts==6.1.0
pillow==10.1.0
pypdf==3.17.0
requests==2.31.0